move dataset manipulation from examples to datasets.utils
