+++ doc/api.rst
@@ -55,4 +55,6 @@ Datasets
     :toctree: generated/
 
     get_mne_sample
+    get_linearly_separable_dataset
+    get_qiskit_dataset
