+++ pyriemann_qiskit/datasets/__init__.py
@@ -1,4 +1,8 @@
-from .utils import get_mne_sample
+from .utils import (get_mne_sample,
+                    get_linearly_separable_dataset,
+                    get_qiskit_dataset)
 
 
-__all__ = ["get_mne_sample"]
+__all__ = ["get_mne_sample", 
+           "get_linearly_separable_dataset", 
+           "get_qiskit_dataset"]
